Deletes transfer histories where In_Spells=1 or In_TransHist=1
Prior to today the transfer history table contained all rows that were
there. It did not have the rows that had been deleted.

This meant in elcid we were loading in rows and not knowing that they
had subsequently deleted (without doing a table scan that is not
practical).

Upstream have kindly changed it so that we now have a 0 in In_Spells or
a 0 in In_TransHist if the spell has been deleted.

This change makes sure we delete those using the encounter slice id, and
then don't create them.