Stops us filtering by mrn before deleting transfer histories
So previously in the transfer history load we would filter by mrn and
then filter by slice id before deleting slice ids. We did this because
it was significantly faster because it used the demographics indexed
hospital number field and then the indexes on foreign keys.

However transfer histories that have had the mrns edited have appeared.
The previous logic meant that the uneditted THs were not deleted. This
change fixes this so we just use slice ID.

This has the effect of making the loader run twice as slowly.