Adds an update_from_dict method that zeros out the original mrn
In the OriginalMRN mixin we have the original_mrn field. This will only
be set by the back end. It is set to the previous MRN of a subrecord
prior to a merge. If it is editted on the front end it can be set to
None as subrecords can only be editted in the front end under their
merged MRN.