Adds the update_from_dict method to the PreviousMRN mixin
The previous_mrn field is updated by the back end when an MRN is merged.
It is used to tell the user that the subrecord was editted under a
previous URL. After the subrecord is saved, we can remove the notice as
its understood the user recognises that it was initially saved in a
previous context.