Ignore initial patient load/external system when updating labtests
When we run the batch ignore the external demographics status of the
patient and ignore the InitialPatientLoad status of the patient.

The external demographics was a hangover from when we used to use the
lab test batch to update demographics. We don't do this any more so it
makes no more send.

The InitialPatientLoad is ignored by all of the other loaders so by
changing lab tests to ignore it we bring lab tests inline with the other
loaders.