Recursively search through the merge comments.
This change changes it so that we recursively iterate through merge
results.

i.e. if MRN 123 is merged into MRN 234 which is merged into 345 which is
merged into 456, we created a MergedMRN object for 123 -> 456.

Prior to this we only went down one level.

This change also changes the definition of a MergedMRN. Prior to this is
what had a quasi relationship where it was inbetween a merge comment and
a row in the master file.

Now it has a 1-1 relationship with a row in the master file.