Removes the ACTIVE/INACTIVE constants from the MergedMRN model.
These are actually just features of the masterfile and are not connected
to the model specifically. Indeed they are not used by any field on the
model.

I've changed the ACTIVE flag to be hardcoded when we load in from the
masterfile and removed the INACTIVE flag as it is never used.