Changes get_active_mrn_and_merged_mrn_data to deal with exceptions
There are 3 sorts of issue we can have with
get_active_mrn_and_merged_mrn_data.

1. We get multiple rows for a single MRN in the masterfile.
This should never happen. In this instance we raise a ValueError because we
should not proceed in this situation.

2. The patient does not exist in the CernerMasterFile.
This should usually not happen. There are cases where we have patients in our
system that are not registered  in the RFH Cerner, these situations need
to be handled by the calling code.

3. There is a typo in the structure of the merge information.
We have instances where there are multiple active patients and instances
where there are only inactive instances. In this case we should allow
the creation of the patient and send us an email so that we can look
into it.