Adds in the boolean rfh_patient argument to get_or_create_patient
Sometimes we deal with patients who are not RFH patients for example
patients from other hospitals who our doctors are advising on.

In these cases they should be able to create a patient but we should not
look for RFH data such as imaging for them.

This argument tells the calling code to not look in upstream databases
for information about the patient.