Adds hardcoded include/ignore lists for models in merge_patient
Hard codes a list of patient/episode related models that should be
copied over. Adds an patient/episode ignore lists in the test.

If a related model is not in either of these lists an error is raised by
the unit test. This forces us to explicitly make sure that related
models are considered when they are added.