Changes the toggle for the services drop down to be the whole div
Changes the toggle for the services menu item drop down to be the whole
div. Prior to this change if you clicked on the edge of the highlighted
box it would not trigger the drop down menu. This changes it so that you
can click anywhere within the list item and the drop down will be
triggered.