Changes the tb review patients list detail to be its own template
We want the TB review patients list to be visible in both the TB menu
and the patient list menu. This does not do this. It does however
abstract the stuff that appears in the main panel into a template that
is included.

This is so that we can include the main panel in another template for
the patient lists.