Changes the link to the TB patients review list from tb.html
Changes it so that there is a button to add/remove the patient from the
list.

Changes it so that if they are on the list it shows in a similar way as
it shows in the teams panel. Clicking on Revew Patients takes you to the
TB Review Patients patient list.