Moves the uib-dropdown-toggle to be on a child element in the service menu
We had an issue that clicking on the dropdown elements in the services
menu did not change the page. This is because the uib-dropdown-toggle
was overriding all links so that they toggled the menu. This resolves
this by moving the dropdown toggle to the <a> which does not encompass
the rest of the menu.