Fixes update_demographics.parse_merge_comments to assign the correct active MRN
It was assigning the MRN that was passed in as the active MRN. This
commit changes it so that it assigns the next_mrn variable. It renames
the mrn variable to initial_mrn as we are dealing with a lot of MRNs so
hopefully this makes it easier to read.