Handle flawed active MRNs in the check_merged_patients cmd
There are a lot of active MRNs with no merge comments, these can be
ignored for the purpose of this command.

There a alot of acitve MRNs where the upstream data is flawed. These
should be check on an individual basis for the time being. All current
flawed patients within elcid have been checked and they are an upstream
data issue and not ours.