Changes add patient pathway to use the new create patient API
In the add patient pathway where possible use get_or_create_patient.

We do get_or_create in the off chance that the patient has been created
in the interim between the search step on the pathway and when we press
save.

We still call super to make sure we save the related subrecords.