Updates the BloodCulturePanelCtrl and BloodCultureIsolate
For isolates we should use the created*/updated*/previous_mrn of the
set.

We don't do this in the to_dict method as this would not display for
new isolates.

After we have updated isolates. We refresh the blood culture set from
the server. This will now have new updated*/previous_mrn* values so
these need to be copied on to the client side blood culture object.