Make sure that IPC status load combines rows correctly
We have multiple rows per patient in the upstream IPC table.
When loading in patients make sure that we combine rows correctly.

The rules are...
If we have a value locally but not upstream, keep out value locally.
If we have a conflict between locally and upstream on a date field,
if its a covid date, use the most recent, otherwise raise a ValueError