Bug fixes the isolates panel for readonly users
As a read only user before we were hiding the icon to add a blood
culture set in blood_culture_set.html. We were not hiding the button to
edit blood cultures.

The edit button does not work for read only users but we should hide it
as well to be consistent.